Backend/requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.6
django-ninja==0.22.2
fpdf==1.7.2
Pillow==10.1.0
pydantic==1.10.13
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
